Add city to location page links from listing
