+++ app/helpers/links_helper.rb
@@ -1,13 +1,4 @@
 module LinksHelper
-  def landing_page_link_or_text(landing_page_criteria, text, match: :exact)
-    lp = match == :exact ? LandingPage.matching(landing_page_criteria) : LandingPage.partially_matching(landing_page_criteria)
-    return tag.span { text } unless lp
-
-    text = match == :exact ? lp.name : landing_page_criteria.values.flatten.first
-    link_text = t("landing_pages.accessible_link_text_html", name: text)
-    govuk_link_to(link_text, landing_page_path(lp.slug), text_colour: true)
-  end
-
   def tracked_link_to(text, href, **kwargs)
     govuk_link_to(text, href, **kwargs.deep_merge(data: {
       controller: "tracked-link",
