+++ app/helpers/vacancies_helper.rb
@@ -70,35 +70,6 @@ module VacanciesHelper
     vacancy.completed_steps.include?(step.to_s)
   end
 
-  def linked_job_roles(vacancy)
-    tag.ul class: "govuk-list" do
-      safe_join(
-        vacancy.job_roles.map do |job_role|
-          tag.li landing_page_link_or_text({ job_roles: [job_role] }, job_role.capitalize)
-        end,
-      )
-    end
-  end
-
-  def linked_subjects(vacancy)
-    vacancy.subjects.map { |subject|
-      landing_page_link_or_text({ subjects: [subject] }, subject, match: :partial)
-    }.join(", ").html_safe
-  end
-
-  def linked_working_patterns(vacancy)
-    tag.ul class: "govuk-list" do
-      safe_join [
-        tag.li do
-          vacancy.working_patterns.map { |working_pattern|
-            landing_page_link_or_text({ working_patterns: [working_pattern] }, working_pattern.capitalize)
-          }.join(", ").html_safe
-        end,
-        tag.li { tag.span(vacancy.working_patterns_details) },
-      ]
-    end
-  end
-
   def map_links(vacancy)
     vacancy.organisations.map do |organisation|
       { text: "#{organisation.name}, #{full_address(organisation)}", url: organisation_url(organisation), id: organisation.id }
